File Upload Widgets

Class and constructor for file upload widgets

Usage

FileUpload(...)

Arguments

...

Any arguments used to initialize the fields of the object

Functions

FileUpload(): The FileUpload constructor function

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> RKernel::ValueWidget -> FileUpload

Public fields

_model_name

Name of the Javascript model in the frontend.

_view_name

Name of the Javascript view in the frontend.

accept

A character string that defineds the accepted file type. If empty, all files are accepted.

multiple

A Boolean traitlet, whether multiple files are accepted.

disabled

A Boolean traitlet, whether the widget is disabled.

icon

A character string, the font-awesome without the 'fa-' prefix.

button_style

The string that describes the button style

style

The button style, an object of class "ButtonStyleClass".

error

A string with an error message, if applicable.

value

The uploaded data.

names

of the uploaded files

Active bindings

names

of the uploaded files

Methods

Public methods

FileUploadClass$new()

FileUploadClass$handle_buffers()

FileUploadClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::ValueWidget$on_change()

Method new()

A generic initializer function

Usage

FileUploadClass$new(description = "Upload", ...)

Arguments

description

The button description

...

Any arguments used to initialize the fields of the object

Method handle_buffers()

Handle buffers in message

Usage

FileUploadClass$handle_buffers(msg)

Arguments

msg

A comm message

Method clone()

The objects of this class are cloneable with this method.

Usage

FileUploadClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.