Progress bars

Classes and constructor functions for progress bars and styling of them

ProgressStyle(...)

IntProgress(value = 0L, min = 0L, max = 100L, ...)

FloatProgress(value = 0L, min = 0L, max = 100L, ...)

Arguments

...

Other arguments.

value

A floating point number, the initial position of the progress bar

min

An floating point number, the minumum value

max

An floating point number, the maximum value

Functions

ProgressStyle(): A constructor for a progress bar style

IntProgress(): A constructor for a progress bar style

FloatProgress(): A constructor for a progress bar style

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DescriptionStyle -> ProgrssStyle

Public fields

_model_name

Name of the Javascript model in the frontend

bar_color

The colour of the progress bar

Methods

Public methods

ProgressStyleClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$initialize()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

Method clone()

The objects of this class are cloneable with this method.

Usage

ProgressStyleClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> RKernel::ValueWidget -> RKernel::BoundedIntWidget -> IntProgress

Public fields

_model_name

Name of the Javascript model in the frontend

_view_name

Name of the Javascript view in the frontend

orientation

Orientation of the progress bar, either "horizontal" or "vertical"

bar_style

General style of the progress bar, either "successs", "info", "warning" or "danger"

style

Styling of the progress bar, an instance of "ProgressStyleClass"

Methods

Public methods

IntProgressClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::ValueWidget$on_change()

RKernel::BoundedIntWidget$initialize()

RKernel::BoundedIntWidget$validate_max()

RKernel::BoundedIntWidget$validate_min()

RKernel::BoundedIntWidget$validate_value()

Method clone()

The objects of this class are cloneable with this method.

Usage

IntProgressClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> RKernel::ValueWidget -> RKernel::BoundedFloatWidget -> FloatProgress

Public fields

_model_name

Name of the Javascript model in the frontend

_view_name

Name of the Javascript view in the frontend

orientation

Orientation of the progress bar, either "horizontal" or "vertical"

bar_style

General style of the progress bar, either "successs", "info", "warning" or "danger"

style

Styling of the progress bar, an instance of "ProgressStyleClass"

Methods

Public methods

FloatProgressClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::ValueWidget$on_change()

RKernel::BoundedFloatWidget$initialize()

RKernel::BoundedFloatWidget$validate_max()

RKernel::BoundedFloatWidget$validate_min()

RKernel::BoundedFloatWidget$validate_value()

Method clone()

The objects of this class are cloneable with this method.

Usage

FloatProgressClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.