The R Session Base Class

An object of this class handles the lower-level communication with an R process. Objects of class RKernelSession inherit from this class.

Super classes

processx::process -> callr::r_session -> RSessionBase

Public fields

prompt

The command prompt

banner

The R startup message used as a session banner in the terminal and info box.

waiting

A logical value, whether the R session is waiting for input.

drop_last_input

A logical value, whether to drop the echo of the last input

last_input

A character string, the last input sent to the R session. Needed to filter out the echo of the input.

Methods

Public methods

RSessionBase$new()

RSessionBase$sleeping()

RSessionBase$send_input()

RSessionBase$read_output()

RSessionBase$receive_output()

RSessionBase$receive_all_output()

RSessionBase$send_receive()

RSessionBase$clone()

Inherited methods

processx::process$as_ps_handle()

processx::process$format()

processx::process$get_cmdline()

processx::process$get_cpu_times()

processx::process$get_error_connection()

processx::process$get_error_file()

processx::process$get_exe()

processx::process$get_exit_status()

processx::process$get_input_connection()

processx::process$get_input_file()

processx::process$get_memory_info()

processx::process$get_name()

processx::process$get_output_connection()

processx::process$get_output_file()

processx::process$get_pid()

processx::process$get_poll_connection()

processx::process$get_result()

processx::process$get_start_time()

processx::process$get_status()

processx::process$get_username()

processx::process$get_wd()

processx::process$has_error_connection()

processx::process$has_input_connection()

processx::process$has_output_connection()

processx::process$has_poll_connection()

processx::process$interrupt()

processx::process$is_alive()

processx::process$is_incomplete_error()

processx::process$is_incomplete_output()

processx::process$is_supervised()

processx::process$kill()

processx::process$kill_tree()

processx::process$poll_io()

processx::process$read_all_error()

processx::process$read_all_error_lines()

processx::process$read_all_output()

processx::process$read_all_output_lines()

processx::process$read_error()

processx::process$read_error_lines()

processx::process$read_output_lines()

processx::process$resume()

processx::process$signal()

processx::process$supervise()

processx::process$suspend()

processx::process$wait()

processx::process$write_input()

callr::r_session$attach()

callr::r_session$call()

callr::r_session$close()

callr::r_session$debug()

callr::r_session$finalize()

callr::r_session$get_running_time()

callr::r_session$get_state()

callr::r_session$poll_process()

callr::r_session$print()

callr::r_session$read()

callr::r_session$run()

callr::r_session$run_with_output()

callr::r_session$traceback()

Method new()

Initialize the object and start the session

Usage

RSessionBase$new(
  options = r_session_options(stdout = "|", stderr = "|", cmdargs = c("--interactive",
    "--no-readline", "--no-save", "--no-restore"), env = c(R_CLI_NUM_COLORS =
    "16777216")),
  prompt = "> "
)

Arguments

options

R session objects, see r_session_options.

prompt

The expected prompt string of the R session

env

A character vector with environment variables for the R process

Method sleeping()

Returns a logical value, indicating whether the R process is sleeping.

Usage

RSessionBase$sleeping()

Method send_input()

Send input text to the R process

Usage

RSessionBase$send_input(text, drop_echo = FALSE)

Arguments

text

A character string

drop_echo

A logical value, whether to drop the echo from stdout.

Method read_output()

Read output from the R session and drop input echo if so requested

Usage

RSessionBase$read_output(n = -1)

Arguments

n

The number of characters to read

Method receive_output()

Poll R process for output and read it

Usage

RSessionBase$receive_output(timeout = 1)

Arguments

timeout

A number, the polling timeout in microseconds

Method receive_all_output()

Receive all output that is available

Usage

RSessionBase$receive_all_output(timeout = 1)

Arguments

timeout

A number, the polling timeout in microseconds

Method send_receive()

Send text to R process and receive all output from the process

Usage

RSessionBase$send_receive(text, timeout = 100)

Arguments

text

A character string

timeout

An integer number, the polling timeout

Method clone()

The objects of this class are cloneable with this method.

Usage

RSessionBase$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.