Sliders

Classes and constructor functions for sliders (integer and floating-point ones)

Usage

IntSlider(value = 0L, min = 0L, max = 100L, ...)

IntRangeSlider(value = c(0L, 50L), min = 0L, max = 100L, ...)

FloatSlider(value = 0, min = 0, max = 100, ...)

FloatRangeSlider(value = c(0, 50), min = 0, max = 100, ...)

FloatLogSlider(value = 1, min = 0, max = 40, base = 10, ...)

Arguments

value

A floating point number, the current value of the slider

min

A floating point number, the minimum value

max

A floating point number, the maximum value

...

Other arguments.

base

A floating point number, the base of the logarithm

Functions

IntSlider(): An integer slider constructor

IntRangeSlider(): An integer range slider constructor

FloatSlider(): A floating point slider constructor

FloatRangeSlider(): A floating point slider range constructor

FloatLogSlider(): A floating point log-slider constructor

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DescriptionStyle -> SliderStyle

Public fields

_model_name

Name of the Javascript frontend model

handle_color

Unicode string, the color of the slider handle

Methods

Public methods

SliderStyleClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$initialize()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

Method clone()

The objects of this class are cloneable with this method.

Usage

SliderStyleClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> RKernel::ValueWidget -> RKernel::BoundedIntWidget -> IntSlider

Public fields

_model_name

Name of the Javascript model in the frontend

_view_name

Name of the Javascript view in the frontend

step

An Integer traitlet, the minimal step size per slider movement

orientation

A Unicode string, either "horizontal" or "vertical"

readout

A logical value, whether the value should be showns (read out)

readout_format

A unicode string, the format specification

continuous_update

A logical value, whether values should be updated as the slider is moved by the user

disabled

A logical value, whether the slider is disabled

style

A SliderStyle widget

behavior

A string that describes the ddragging behavior.

Methods

Public methods

IntSliderClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::ValueWidget$on_change()

RKernel::BoundedIntWidget$initialize()

RKernel::BoundedIntWidget$validate_max()

RKernel::BoundedIntWidget$validate_min()

RKernel::BoundedIntWidget$validate_value()

Method clone()

The objects of this class are cloneable with this method.

Usage

IntSliderClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> RKernel::ValueWidget -> RKernel::BoundedIntRangeWidget -> IntRangeSlider

Public fields

_model_name

Name of the Javascript model in the frontend

_view_name

Name of the Javascript view in the frontend

step

An Integer traitlet, the minimal step size per slider movement

orientation

A Unicode string, either "horizontal" or "vertical"

readout

A logical value, whether the value should be showns (read out)

readout_format

A logical value, whether values should be updated as the slider is moved by the user

continuous_update

A logical value, whether values should be updated as the slider is moved by the user

disabled

A logical value, whether the slider is disabled

style

A SliderStyle widget

behavior

A string that describes the ddragging behavior.

Methods

Public methods

IntRangeSliderClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::ValueWidget$on_change()

RKernel::BoundedIntRangeWidget$initialize()

RKernel::BoundedIntRangeWidget$validate_max()

RKernel::BoundedIntRangeWidget$validate_min()

RKernel::BoundedIntRangeWidget$validate_value()

Method clone()

The objects of this class are cloneable with this method.

Usage

IntRangeSliderClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> RKernel::ValueWidget -> RKernel::BoundedFloatWidget -> FloatSlider

Public fields

_model_name

Name of the Javascript model in the frontend

_view_name

Name of the Javascript view in the frontend

step

A Float traitlet, the minimal step size per slider movement

orientation

A Unicode string, either "horizontal" or "vertical"

readout

A logical value, whether the value should be showns (read out)

readout_format

A logical value, whether values should be updated as the slider is moved by the user

continuous_update

A logical value, whether values should be updated as the slider is moved by the user

disabled

A logical value, whether the slider is disabled

style

A SliderStyle widget

behavior

A string that describes the ddragging behavior.

Methods

Public methods

FloatSliderClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::ValueWidget$on_change()

RKernel::BoundedFloatWidget$initialize()

RKernel::BoundedFloatWidget$validate_max()

RKernel::BoundedFloatWidget$validate_min()

RKernel::BoundedFloatWidget$validate_value()

Method clone()

The objects of this class are cloneable with this method.

Usage

FloatSliderClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> RKernel::ValueWidget -> RKernel::BoundedFloatRangeWidget -> FloatRangeSlider

Public fields

_model_name

Name of the Javascript model in the frontend

_view_name

Name of the Javascript view in the frontend

step

A Float traitlet, the minimal step size per slider movement

orientation

A Unicode string, either "horizontal" or "vertical"

readout

A logical value, whether the value should be showns (read out)

readout_format

A logical value, whether values should be updated as the slider is moved by the user

continuous_update

A logical value, whether values should be updated as the slider is moved by the user

disabled

A logical value, whether the slider is disabled

style

A SliderStyle widget

behavior

A string that describes the ddragging behavior.

Methods

Public methods

FloatRangeSliderClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::ValueWidget$on_change()

RKernel::BoundedFloatRangeWidget$initialize()

RKernel::BoundedFloatRangeWidget$validate_max()

RKernel::BoundedFloatRangeWidget$validate_min()

RKernel::BoundedFloatRangeWidget$validate_value()

Method clone()

The objects of this class are cloneable with this method.

Usage

FloatRangeSliderClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> RKernel::ValueWidget -> RKernel::BoundedLogFloatWidget -> FloatLogSlider

Public fields

_model_name

Name of the Javascript model in the frontend

_view_name

Name of the Javascript view in the frontend

step

A Float traitlet, the minimal step size per slider movement

orientation

A Unicode string, either "horizontal" or "vertical"

readout

A logical value, whether the value should be showns (read out)

readout_format

A logical value, whether values should be updated as the slider is moved by the user

continuous_update

A logical value, whether values should be updated as the slider is moved by the user

disabled

A Boolean traitlet, whether the slider is disabled

base

A Float traitlet, the base of the logarithm

style

A SliderStyle widget

behavior

A string that describes the ddragging behavior.

Methods

Public methods

FloatLogSliderClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::ValueWidget$on_change()

RKernel::BoundedLogFloatWidget$initialize()

RKernel::BoundedLogFloatWidget$validate_max()

RKernel::BoundedLogFloatWidget$validate_min()

RKernel::BoundedLogFloatWidget$validate_value()

Method clone()

The objects of this class are cloneable with this method.

Usage

FloatLogSliderClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.