Tobble-Button Widgets

A constructor function and a class to create toggle-button widgets

Usage

ToggleButtonStyle(...)

ToggleButton(...)

Arguments

...

Arguments passed to the inializer

Functions

ToggleButtonStyle(): The constructor for Togglebuttons styles

ToggleButton(): A toggle-button constructor

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DescriptionStyle -> ToggleButtonStyle

Public fields

_model_name

Name of the Javascript model in the frontend

font_family

The font family

font_size

The font size

font_style

The font style

font_variant

The font variant

font_weight

The font weight

text_color

The text color

text_decoration

The text decoration

required_version

Minimum required ipywidgets version in which the current widget class is supported.

Methods

Public methods

ToggleButtonStyleClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$initialize()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

Method clone()

The objects of this class are cloneable with this method.

Usage

ToggleButtonStyleClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> RKernel::ValueWidget -> ToggleButton

Public fields

_model_name

Name of the Javascript model in the frontend

_view_name

Name of the Javascript model view in the frontend

value

Boolean, whether the box is checked

tooltip

A tooltip

description

A button description

disabled

Boolean, whether the button is disabled

icon

An icon (a fontawesome icon name)

button_style

The string that describes the button style

style

The toggle button style, an object of class "ToggleButtonStyleClass"

Methods

Public methods

ToggleButtonClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::ValueWidget$initialize()

RKernel::ValueWidget$on_change()

Method clone()

The objects of this class are cloneable with this method.

Usage

ToggleButtonClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.