A HTML version of 'ls.str()'

This function is deprecated. Use envBrowser instead.

ls_str(pos = -1, name, envir, all.names = FALSE, pattern, mode = "any")

Arguments

pos

integer indicating search path position, or -1 for the current environment.

name

an optional name indicating search path position, see ls.

envir

the environment to look into

all.names

logical value, if FALSE objects with names that start with a dot are ignored

pattern

a character string, the pattern of the names of the objects to show

mode

a character string, the mode of the objects to be shown

See also