Support for AnyWidgets

Support for AnyWidgets

Support for AnyWidgets

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> AnyWidget

Public fields

_model_module

Name of the Javascript module with the model

_model_module_version

Version of the module where the model is defined

_model_name

Name of the Javascript model in the frontend

_view_name

Name of the Javascript model view in the frontend

_view_module

Name of the module where the view is defined

_view_module_version

Version of the module where the view is defined

_anywidget_id

A Unicode trait

_css

A Unicode trait with optional CSS code

_esm

A Unicode trait with optional Javascript code

Methods

Public methods

AnyWidgetClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$initialize()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

Method clone()

The objects of this class are cloneable with this method.

Usage

AnyWidgetClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.