Widgets for Bounded Floating Point Numbers on a logarithmic scale

An R6 class and a constructor function for the creation of widgets that can be used to manipulate floating point numbers that are bounded within an interval on an logarithmic scale

BoundedLogFloatWidget(value, min, max, base, ...)

Arguments

value

The floating point value.

min

The lower bound of the interval.

max

The upper bound of the interval.

base

The base of the logarithm.

...

Other arguments, passed to the superclass initializer.

Details

The function BoundedLogFloatWidget creates objects of the R6 Class "BoundedLogFloatWidgetClass", which in turn have the S3 class attribute "BoundedLogFloatWidget"

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> RKernel::ValueWidget -> BoundedLogFloatWidget

Public fields

value

A Float traitlet.

min

A Float traitlet, the minimum allowed value.

max

A Float traitlet, the maximum allowed value.

base

A Float traitlet, the logarithmic base.

Methods

Public methods

BoundedLogFloatWidgetClass$validate_value()

BoundedLogFloatWidgetClass$validate_min()

BoundedLogFloatWidgetClass$validate_max()

BoundedLogFloatWidgetClass$new()

BoundedLogFloatWidgetClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::ValueWidget$on_change()

Method validate_value()

Validate the "value" after assignment.

Usage

BoundedLogFloatWidgetClass$validate_value(value)

Arguments

value

A value, should be numeric.

Method validate_min()

Validate the "min" field after assignment.

Usage

BoundedLogFloatWidgetClass$validate_min(min)

Arguments

min

A minimum value, should be numeric.

Method validate_max()

Validate the "max" field after assignment.

Usage

BoundedLogFloatWidgetClass$validate_max(max)

Arguments

max

A maximum value, should be numeric.

Method new()

Initialize an object.

Usage

BoundedLogFloatWidgetClass$new(value, min, max, ...)

Arguments

value

The floating point value.

min

The lower bound of the interval.

max

The upper bound of the interval.

...

Other arguments, passed to the superclass initializer.

Method clone()

The objects of this class are cloneable with this method.

Usage

BoundedLogFloatWidgetClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.