Box Containers

Classes and constructor functions to create box container widgets

Usage

Box(..., layout = NULL)

HBox(..., layout = NULL, wrap = FALSE)

VBox(..., layout = NULL, wrap = FALSE)

GridBox(..., layout = NULL)

Arguments

...

Arguments passed to the superclass constructor

wrap

Logical value, whether lines of widgets should be wrapped?

Details

VBox creates vertical boxes, HBox creates horizontal boxes, GridBox creates a grid box. In a vertical box widgets are arranged one below the other, in a horizontal box widgets are arranged side-by-side, in a grid box widget are arraned in a grid.

Functions

Box(): A baseline box constructor

HBox(): A horizontal box constructor

VBox(): A vertical box constructor

GridBox(): A grid box constructor

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> Box

Public fields

_model_module

Name of the Javascript module with the model

_model_module_version

Version of the module where the model is defined

_model_name

Name of the Javascript model in the frontend

_view_name

Name of the Javascript model view in the frontend

_view_module

Name of the module where the view is defined

_view_module_version

Version of the module where the view is defined

children

A generic vector with the widgets in the container

box_style

The string that describes the button style

Methods

Public methods

BoxClass$new()

BoxClass$notify_children_displayed()

BoxClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

Method new()

An initializer function

Usage

BoxClass$new(children = list(), ...)

Arguments

children

A list of widgets

...

Other arguments, passed to the superclass method

Method notify_children_displayed()

Notifies children that they are displayed

Usage

BoxClass$notify_children_displayed()

Method clone()

The objects of this class are cloneable with this method.

Usage

BoxClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::Box -> HBox

Public fields

_model_name

Name of the Javascript model in the frontend

_view_name

Name of the Javascript model view in the frontend

Methods

Public methods

HBoxClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::Box$initialize()

RKernel::Box$notify_children_displayed()

Method clone()

The objects of this class are cloneable with this method.

Usage

HBoxClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::Box -> VBox

Public fields

_model_name

Name of the Javascript model in the frontend

_view_name

Name of the Javascript model view in the frontend

Methods

Public methods

VBoxClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::Box$initialize()

RKernel::Box$notify_children_displayed()

Method clone()

The objects of this class are cloneable with this method.

Usage

VBoxClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::Box -> GridBox

Public fields

_model_name

Name of the Javascript model in the frontend

_view_name

Name of the Javascript model view in the frontend

Methods

Public methods

GridBoxClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::Box$initialize()

RKernel::Box$notify_children_displayed()

Method clone()

The objects of this class are cloneable with this method.

Usage

GridBoxClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.