A Base Class for DOM Widgets

This is a base class for all widgets that are supposed to be part of the document object model

Usage

DOMWidget(...)

Arguments

...

Arguments passed to the inializer

Functions

DOMWidget(): The DOM widget constructor function

Super classes

RKernel::HasTraits -> RKernel::Widget -> DOMWidget

Public fields

_model_module

Name of the Javascript module with the model

_model_module_version

Version of the module where the model is defined

_model_name

Name of the Javascript model in the frontend

_dom_classes

A set of character strings that indicate the DOM classes the widget is assigned to

layout

The layout, a "LayoutClass" Widget

Methods

Public methods

DOMWidgetClass$add_class()

DOMWidgetClass$remove_class()

DOMWidgetClass$has_class()

DOMWidgetClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$initialize()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

Method add_class()

Add a class attribute to the DOM element

Usage

DOMWidgetClass$add_class(className)

Arguments

className

Name of the class attribute

Method remove_class()

Remove a class attribute to the DOM element

Usage

DOMWidgetClass$remove_class(className)

Arguments

className

Name of the class attribute

Method has_class()

Check whether the DOM element has a class attribute

Usage

DOMWidgetClass$has_class(className)

Arguments

className

Name of the class attribute

Method clone()

The objects of this class are cloneable with this method.

Usage

DOMWidgetClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.