Dictionary Traitlets

A class and a constructor of dictionary trait(let)s. These are lists with unique element names.

Usage

Dict(...)

Arguments

...

Arguments that are passed to the initialize method of 'DictClass'

Super classes

RKernel::Trait -> RKernel::List -> Dict

Methods

Public methods

DictClass$validator()

DictClass$clone()

Inherited methods

RKernel::Trait$get()

RKernel::Trait$initialize()

RKernel::Trait$set()

Method validator()

A function that checks the validity of an assigned value, i.e. whether the assigned value is a list with unique names

Usage

DictClass$validator(value)

Arguments

value

A value to be assigned as the traitlet value

Method clone()

The objects of this class are cloneable with this method.

Usage

DictClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.