Widget Layout Templates

R6 classes and constructor functions for widget layout templates

Usage

AppLayout(...)

GridspecLayout(...)

# S3 method for GridspecLayout
[(x, i, j, ..., drop = TRUE)

# S3 method for GridspecLayout
[(x, i, j) <- value

TwoByTwoLayout(...)

Arguments

...

Arguments used to initialize the fields

x

A GridspecLayout object

i

Integer value(s) referring to the row(s)

j

Integer value(s) referring to the column(s)

drop

Logical, whether the result is a widget or a list with one element if both i an j select a single element

value

One or more widgets put at the idicated positions in the grid

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::Box -> RKernel::GridBox -> TemplateBase

Public fields

grid_gap

The grid-gap CSS attribute

justify_content

The justify-content CSS attribute

align_items

The align-items CSS attribute

width

The width CSS attribute

height

The height CSS attribute

Methods

Public methods

TemplateBaseClass$new()

TemplateBaseClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::Box$notify_children_displayed()

Method new()

Initializer

Usage

TemplateBaseClass$new(...)

Arguments

...

Arguments used to initialize the fields

Method clone()

The objects of this class are cloneable with this method.

Usage

TemplateBaseClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::Box -> RKernel::GridBox -> RKernel::TemplateBase -> AppLayout

Public fields

header

Widget that appears in the header section

footer

Widget that appears in the footer section

left_sidebar

Widget that appears as left sidebar

right_sidebar

Widget that appears as right sidebar

center

Widget that appears in the center section

pane_widths

Unicode string with CSS widths for the app panes

pane_heights

Unicode string with CSS heights for the app panes

merge

Boolean, whether space of missing widgets chould be merged

Methods

Public methods

AppLayoutClass$new()

AppLayoutClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::Box$notify_children_displayed()

Method new()

Initializer method

Usage

AppLayoutClass$new(...)

Arguments

...

Arguments, passed on to the superclass initializer

Method clone()

The objects of this class are cloneable with this method.

Usage

AppLayoutClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::Box -> RKernel::GridBox -> RKernel::TemplateBase -> GridspecLayout

Methods

Public methods

GridspecLayoutClass$new()

GridspecLayoutClass$set_item()

GridspecLayoutClass$get_item()

GridspecLayoutClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::Box$notify_children_displayed()

Method new()

Initializer function

Usage

GridspecLayoutClass$new(nrow, ncol, ...)

Arguments

nrow

A positive integer, the number of rows

ncol

A positive integer, the number of columns

...

Other arguments, passed to the superclass initializer

Method set_item()

Set widget in grid cells

Usage

GridspecLayoutClass$set_item(i, j, value)

Arguments

i

The rows into which the widget is to be placed

j

The columns into which the widget is to be placed

value

A widget

Method get_item()

Get widget from grid cells

Usage

GridspecLayoutClass$get_item(i, j)

Arguments

i

The rows where the widget is located

j

The columns where the widget is located

Method clone()

The objects of this class are cloneable with this method.

Usage

GridspecLayoutClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::Box -> RKernel::GridBox -> RKernel::TemplateBase -> TwoByTwoLayout

Public fields

top_left

Widget that appears on the top left

top_right

Widget that appears on the top right

bottom_left

Widget that appears on the bottom left

bottom_right

Widget that appears on the bottom right

merge

Boolean, whether space of missing widgets chould be merged

Methods

Public methods

TwoByTwoLayoutClass$new()

TwoByTwoLayoutClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::Box$notify_children_displayed()

Method new()

Initializer method

Usage

TwoByTwoLayoutClass$new(...)

Arguments

...

Arguments, passed on to the superclass initializer

Method clone()

The objects of this class are cloneable with this method.

Usage

TwoByTwoLayoutClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.