A Base Class for Traits that are R6 Objects

A Base Class for Traits that are R6 Objects

A Base Class for Traits that are R6 Objects

Super class

RKernel::Trait -> R6Trait

Public fields

value

An R6 object

class

The R6 class of value

Methods

Public methods

R6TraitClass$validator()

R6TraitClass$new()

R6TraitClass$clone()

Inherited methods

RKernel::Trait$get()

RKernel::Trait$set()

Method validator()

Checks wether value has the corret class

Usage

R6TraitClass$validator(value)

Arguments

value

A value about to be assigned to the trait.

Method new()

Initialize an object

Usage

R6TraitClass$new(Class, ...)

Arguments

Class

Class of the object

...

Values used for initialization

Method clone()

The objects of this class are cloneable with this method.

Usage

R6TraitClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.