The KernelSession Class

An object of this class handles the communication with an R process. There is usually one main session, but there may also sessions for detached code cells.

Super classes

processx::process -> callr::r_session -> RKernel::RSessionBase -> RKernelSession

Public fields

help_port

The port number of HTML help.

yield

An optional function to service kernel requests

kernel

The reference to the controlling kernel object

Methods

Public methods

RKernelSession$connect()

RKernelSession$setup()

RKernelSession$start_graphics()

RKernelSession$dev_new()

RKernelSession$graphics_details()

RKernelSession$clone()

Inherited methods

processx::process$as_ps_handle()

processx::process$format()

processx::process$get_cmdline()

processx::process$get_cpu_times()

processx::process$get_error_connection()

processx::process$get_error_file()

processx::process$get_exe()

processx::process$get_exit_status()

processx::process$get_input_connection()

processx::process$get_input_file()

processx::process$get_memory_info()

processx::process$get_name()

processx::process$get_output_connection()

processx::process$get_output_file()

processx::process$get_pid()

processx::process$get_poll_connection()

processx::process$get_result()

processx::process$get_start_time()

processx::process$get_status()

processx::process$get_username()

processx::process$get_wd()

processx::process$has_error_connection()

processx::process$has_input_connection()

processx::process$has_output_connection()

processx::process$has_poll_connection()

processx::process$interrupt()

processx::process$is_alive()

processx::process$is_incomplete_error()

processx::process$is_incomplete_output()

processx::process$is_supervised()

processx::process$kill()

processx::process$kill_tree()

processx::process$poll_io()

processx::process$read_all_error()

processx::process$read_all_error_lines()

processx::process$read_all_output()

processx::process$read_all_output_lines()

processx::process$read_error()

processx::process$read_error_lines()

processx::process$read_output_lines()

processx::process$resume()

processx::process$signal()

processx::process$supervise()

processx::process$suspend()

processx::process$wait()

processx::process$write_input()

callr::r_session$attach()

callr::r_session$call()

callr::r_session$close()

callr::r_session$debug()

callr::r_session$finalize()

callr::r_session$get_running_time()

callr::r_session$get_state()

callr::r_session$poll_process()

callr::r_session$print()

callr::r_session$read()

callr::r_session$run()

callr::r_session$run_with_output()

callr::r_session$traceback()

RKernel::RSessionBase$initialize()

RKernel::RSessionBase$read_output()

RKernel::RSessionBase$receive_all_output()

RKernel::RSessionBase$receive_output()

RKernel::RSessionBase$send_input()

RKernel::RSessionBase$send_receive()

RKernel::RSessionBase$sleeping()

Method connect()

Connect the session with the main Kernel object

Usage

RKernelSession$connect(yield, kernel)

Arguments

yield

The function that is called when the session objects yields back control to the kernel object, for example in order to wait for comm message when a widget is active.

kernel

The main kernel objectd

Method setup()

Set up the R session, by installing hooks etc.

Usage

RKernelSession$setup()

Method start_graphics()

Initialize graphics, start device and return details

Usage

RKernelSession$start_graphics()

Method dev_new()

Start a httpgd device and return the graphics details.

Usage

RKernelSession$dev_new()

Method graphics_details()

Return the graphics details.

Usage

RKernelSession$graphics_details()

Method clone()

The objects of this class are cloneable with this method.

Usage

RKernelSession$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.