Tabs and Accordions

Classes and constructor functions for tab and accordion widgets

Usage

Accordion(...)

Tab(...)

Stack(...)

Arguments

...

Arguments passed to the superclass constructor

Functions

Accordion(): The constructor function for accordion widgets.

Tab(): The construction function for accordion widgets.

Stack(): The construction function for accordion widgets.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::Box -> Box

Public fields

_titles

A dictionary of strings, for internal use only

titles

A dictionary of strings, exposed since ipywidgets 8.

selected_index

An integer, the field currently selected.

Methods

Public methods

SelectionContainerClass$validate_index()

SelectionContainerClass$set_title()

SelectionContainerClass$get_title()

SelectionContainerClass$new()

SelectionContainerClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::Box$notify_children_displayed()

Method validate_index()

Validate the index, i.e. check whether it is within range.

Usage

SelectionContainerClass$validate_index(index)

Arguments

index

An integer number.

Method set_title()

Set the title of one of the elements.

Usage

SelectionContainerClass$set_title(index, title)

Arguments

index

The index number of the element to be changed.

title

A character string, the intended title.

Method get_title()

Get the title of one of the elements.

Usage

SelectionContainerClass$get_title(index)

Arguments

index

The index number of the element to be enquired.

Method new()

An initializer function

Usage

SelectionContainerClass$new(children = list(), ...)

Arguments

children

A list of widgets

...

Other arguments, passed to the superclass method

Method clone()

The objects of this class are cloneable with this method.

Usage

SelectionContainerClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::Box -> Accordion

Public fields

_model_name

Name of the Javascript model in the frontend.

_view_name

Name of the Javascript view in the frontend.

Methods

Public methods

AccordionClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::Box$notify_children_displayed()

RKernel::Box$get_title()

RKernel::Box$initialize()

RKernel::Box$set_title()

RKernel::Box$validate_index()

Method clone()

The objects of this class are cloneable with this method.

Usage

AccordionClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::Box -> Tab

Public fields

_model_name

Name of the Javascript model in the frontend.

_view_name

Name of the Javascript view in the frontend.

Methods

Public methods

TabClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::Box$notify_children_displayed()

RKernel::Box$get_title()

RKernel::Box$initialize()

RKernel::Box$set_title()

RKernel::Box$validate_index()

Method clone()

The objects of this class are cloneable with this method.

Usage

TabClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::Box -> Stack

Public fields

_model_name

Name of the Javascript model in the frontend.

_view_name

Name of the Javascript view in the frontend.

required_version

Minimum required ipywidgets version in which the current widget class is supported.

Methods

Public methods

StackClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::Box$notify_children_displayed()

RKernel::Box$get_title()

RKernel::Box$initialize()

RKernel::Box$set_title()

RKernel::Box$validate_index()

Method clone()

The objects of this class are cloneable with this method.

Usage

StackClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.