Selection widgets

Classes and constructors for selection widgets, i.e. dropdowns, listboxes etc.

Usage

SelectionWidget(options, value, ...)

Dropdown(options, value, ...)

RadioButtons(options, value, ...)

ListBox(options, value, ...)

ToggleButtons(options, value, ...)

SelectionSlider(options, value, ...)

MultipleSelectionWidget(options, value, ...)

ListBoxMultiple(options, value, ...)

SelectionRangeSlider(options, value, ...)

Arguments

options

A named vector or a vector coerceable into a character vector.

value

A trait.

...

Any other arguments, ignored.

Functions

SelectionWidget(): A constructor function for selection widgets.

Dropdown(): The construction function for dropdown widgets.

RadioButtons(): The construction function for radiobuttons widgets.

ListBox(): The construction function for listbox-selection widgets.

ToggleButtons(): The construction function for togglebuttons widgets.

SelectionSlider(): The construction function for listbox widgets with multiple selections.

MultipleSelectionWidget(): The construction function for multiple-selection widgets.

ListBoxMultiple(): The construction function for listbox widgets with multiple selections.

SelectionRangeSlider(): The construction function for listbox widgets with multiple selections.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> SelectionWidget

Public fields

_options_labels

A unicode vector of option labels.

index

An integer that refers to currently selected item.

Active bindings

value

The selected option

Methods

Public methods

SelectionWidgetClass$validate_index()

SelectionWidgetClass$new()

SelectionWidgetClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

Method validate_index()

Validate an index argument.

Usage

SelectionWidgetClass$validate_index(index)

Arguments

index

The index to be checked.

Method new()

Initialiser

Usage

SelectionWidgetClass$new(options, value, ...)

Arguments

options

A named vector or a vector coerceable into a character vector.

value

Name of a selectable option.

...

Any other arguments, ignored.

Method clone()

The objects of this class are cloneable with this method.

Usage

SelectionWidgetClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> RKernel::SelectionWidget -> Dropdown

Public fields

_model_name

Name of the Javascript model in the frontend.

_view_name

Name of the Javascript view in the frontend.

Methods

Public methods

DropdownClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::SelectionWidget$initialize()

RKernel::SelectionWidget$validate_index()

Method clone()

The objects of this class are cloneable with this method.

Usage

DropdownClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> RKernel::SelectionWidget -> RadioButtons

Public fields

_model_name

Name of the Javascript model in the frontend.

_view_name

Name of the Javascript view in the frontend.

Methods

Public methods

RadioButtonsClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::SelectionWidget$initialize()

RKernel::SelectionWidget$validate_index()

Method clone()

The objects of this class are cloneable with this method.

Usage

RadioButtonsClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> RKernel::SelectionWidget -> ListboxSelect

Public fields

_model_name

Name of the Javascript model in the frontend.

_view_name

Name of the Javascript view in the frontend.

rows

An integer, the number of rows.

Methods

Public methods

ListboxSelectClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::SelectionWidget$initialize()

RKernel::SelectionWidget$validate_index()

Method clone()

The objects of this class are cloneable with this method.

Usage

ListboxSelectClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DescriptionStyle -> ToggleButtonsStyle

Public fields

_model_name

Name of the Javascript model in the frontend.

button_width

A unicode string, the width in CSS language

font_weight

A unicode string, the font weight in CSS language

Methods

Public methods

ToggleButtonsStyleClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$initialize()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

Method clone()

The objects of this class are cloneable with this method.

Usage

ToggleButtonsStyleClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> RKernel::SelectionWidget -> ToggleButtons

Public fields

_model_name

Name of the Javascript model in the frontend.

_view_name

Name of the Javascript view in the frontend.

tooltips

A unicode vector with tooltips.

icons

A unicode vector with icon specs.

style

A TobbleButtonStyle widget

button_style

A character string, one of "primary", "success", "info", "warning", "danger", or the empty string.

Methods

Public methods

ToggleButtonsClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::SelectionWidget$initialize()

RKernel::SelectionWidget$validate_index()

Method clone()

The objects of this class are cloneable with this method.

Usage

ToggleButtonsClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> RKernel::SelectionWidget -> SelectionSlider

Public fields

_model_name

Name of the Javascript model in the frontend.

_view_name

Name of the Javascript view in the frontend.

orientation

A Unicode string, either "horizontal" or "vertical"

readout

A logical value, whether the value should be showns (read out)

continuous_update

A logical value, whether values should be updated as the slider is moved by the user

style

A SliderStyle widget

behavior

A string that describes the ddragging behavior.

Methods

Public methods

SelectionSliderClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::SelectionWidget$initialize()

RKernel::SelectionWidget$validate_index()

Method clone()

The objects of this class are cloneable with this method.

Usage

SelectionSliderClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> MultipleSelectionWidget

Public fields

_options_labels

A unicode string vector with labels

index

An integer vector of indices of currenlty selected elements.

Active bindings

value

The selected option

Methods

Public methods

MultipleSelectionWidgetClass$validate_index()

MultipleSelectionWidgetClass$new()

MultipleSelectionWidgetClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

Method validate_index()

Validate an index.

Usage

MultipleSelectionWidgetClass$validate_index(index)

Arguments

index

An index, the index to be checked.

Method new()

Initialiser

Usage

MultipleSelectionWidgetClass$new(options, value, ...)

Arguments

options

A named vector or a vector coerceable into a character vector.

value

Names of selectable options.

...

Any other arguments, ignored.

Method clone()

The objects of this class are cloneable with this method.

Usage

MultipleSelectionWidgetClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> RKernel::MultipleSelectionWidget -> ListboxSelectMultiple

Public fields

_model_name

Name of the Javascript model in the frontend.

_view_name

Name of the Javascript view in the frontend.

rows

An integer, the number of rows.

Methods

Public methods

ListboxSelectMultipleClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::MultipleSelectionWidget$initialize()

RKernel::MultipleSelectionWidget$validate_index()

Method clone()

The objects of this class are cloneable with this method.

Usage

ListboxSelectMultipleClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> RKernel::MultipleSelectionWidget -> SelectionRangeSlider

Public fields

_model_name

Name of the Javascript model in the frontend.

_view_name

Name of the Javascript view in the frontend.

orientation

A Unicode string, either "horizontal" or "vertical"

readout

A logical value, whether the value should be showns (read out)

continuous_update

A logical value, whether values should be updated as the slider is moved by the user

style

A SliderStyle widget

behavior

A string that describes the ddragging behavior.

Methods

Public methods

SelectionRangeSliderClass$validate_index()

SelectionRangeSliderClass$new()

SelectionRangeSliderClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

Method validate_index()

Validate an index.

Usage

SelectionRangeSliderClass$validate_index(index)

Arguments

index

An index, the index to be checked.

Method new()

Initialiser

Usage

SelectionRangeSliderClass$new(options, value, ...)

Arguments

options

A named vector or a vector coerceable into a character vector.

value

Names of selectable options.

...

Any other arguments, ignored.

Method clone()

The objects of this class are cloneable with this method.

Usage

SelectionRangeSliderClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.