Sidecar widgets

Sidecar widgets - work only with Jupyter Lab

Usage

Sidecar(...)

Arguments

...

Arguments passed to the inializer

Functions

Sidecar(): A constructor for sidebar widgets

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::OutputWidget -> Sidecar

Public fields

_model_name

Name of the Javascript model in the frontend

_model_module

Name of the Javascript frontend module

_model_module_version

Version of the Javascript frontend module

_view_name

Name of the Javascript view in the frontend

_view_module

Name of the Javascript frontend view module

_view_module_version

Version of the the Javascript view module

title

A unicode string, the title of the widget.

anchor

A string that specifies where the widget s to appear: one of "split-right", "split-left", "split-top", "split-bottom", "tab-before", "tab-after", or "right".

Methods

Public methods

SidecarClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::OutputWidget$cat()

RKernel::OutputWidget$clear_output()

RKernel::OutputWidget$display()

RKernel::OutputWidget$do()

RKernel::OutputWidget$eval()

RKernel::OutputWidget$evaluate()

RKernel::OutputWidget$initialize()

RKernel::OutputWidget$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

SidecarClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.