String widgets

Classes and constructor functions for string-related widgets (text areas etc.)

LabelStyle(...)

TextStyle(...)

HTMLStyle(...)

HTMLMathStyle(...)

StringWidget(value = character(0), ...)

HTML(value = character(0), ...)

HTMLMath(value = character(0), ...)

Label(value = character(0), ...)

Textarea(value = character(0), ...)

TextWidget(value = character(0), ...)

PasswordWidget(value = character(0), ...)

Combobox(value = character(0), ...)

Arguments

...

Arguments passed to the inializer

value

A character vector

Functions

LabelStyle(): The constructor for Label styles

TextStyle(): The constructor for Text styles

HTMLStyle(): The constructor for HTML styles

HTMLMathStyle(): The constructor for HTMLMath styles

StringWidget(): A constructor for string widgets

HTML(): A constructor for HTML widgets

HTMLMath(): A constructor for HTML widgets with math

Label(): A constructor for label widgets

Textarea(): A constructor for text area widgets

TextWidget(): A constructor for text field widgets

PasswordWidget(): A constructor for password entry widgets

Combobox(): A constructor for combo boxes

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DescriptionStyle -> StringStyle

Public fields

_model_name

Name of the Javascript model in the frontend

background

The background color

font_size

The font size

text_color

The text color

required_version

Minimum required ipywidgets version in which the current widget class is supported.

Methods

Public methods

StringStyleClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$initialize()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

Method clone()

The objects of this class are cloneable with this method.

Usage

StringStyleClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DescriptionStyle -> RKernel::StringStyle -> LabelStyle

Public fields

_model_name

Name of the Javascript model in the frontend

font_family

The font family

font_style

The font style

font_variant

The font variant

font_weight

The font weight

text_decoration

The text decoration

required_version

Minimum required ipywidgets version in which the current widget class is supported.

Methods

Public methods

LabelStyleClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$initialize()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

Method clone()

The objects of this class are cloneable with this method.

Usage

LabelStyleClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DescriptionStyle -> RKernel::StringStyle -> TextStyle

Public fields

_model_name

Name of the Javascript model in the frontend

required_version

Minimum required ipywidgets version in which the current widget class is supported.

Methods

Public methods

TextStyleClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$initialize()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

Method clone()

The objects of this class are cloneable with this method.

Usage

TextStyleClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DescriptionStyle -> RKernel::StringStyle -> HTMLStyle

Public fields

_model_name

Name of the Javascript model in the frontend

required_version

Minimum required ipywidgets version in which the current widget class is supported.

Methods

Public methods

HTMLStyleClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$initialize()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

Method clone()

The objects of this class are cloneable with this method.

Usage

HTMLStyleClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DescriptionStyle -> RKernel::StringStyle -> HTMLMathStyle

Public fields

_model_name

Name of the Javascript model in the frontend

required_version

Minimum required ipywidgets version in which the current widget class is supported.

Methods

Public methods

HTMLMathStyleClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$initialize()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

Method clone()

The objects of this class are cloneable with this method.

Usage

HTMLMathStyleClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> RKernel::ValueWidget -> StringWidget

Public fields

_model_name

Name of the Javascript model in the frontend

value

A unicode vector

placeholder

A placeholder character

Methods

Public methods

StringWidgetClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::ValueWidget$initialize()

RKernel::ValueWidget$on_change()

Method clone()

The objects of this class are cloneable with this method.

Usage

StringWidgetClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> RKernel::ValueWidget -> RKernel::StringWidget -> HTML

Public fields

_model_name

Name of the Javascript model in the frontend

_view_name

Name of the Javascript view in the frontend

style

The HTML style, an object of class "HTMLStyleClass"

Methods

Public methods

HTMLClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::ValueWidget$initialize()

RKernel::ValueWidget$on_change()

Method clone()

The objects of this class are cloneable with this method.

Usage

HTMLClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> RKernel::ValueWidget -> RKernel::StringWidget -> HTMLMath

Public fields

_model_name

Name of the Javascript model in the frontend

_view_name

Name of the Javascript view in the frontend

style

The HTMLMath style, an object of class "HTMLMathStyleClass"

Methods

Public methods

HTMLMathClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::ValueWidget$initialize()

RKernel::ValueWidget$on_change()

Method clone()

The objects of this class are cloneable with this method.

Usage

HTMLMathClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> RKernel::ValueWidget -> RKernel::StringWidget -> Label

Public fields

_model_name

Name of the Javascript model in the frontend

_view_name

Name of the Javascript view in the frontend

style

The Label style, an object of class "LabelStyleClass"

Methods

Public methods

LabelClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::ValueWidget$initialize()

RKernel::ValueWidget$on_change()

Method clone()

The objects of this class are cloneable with this method.

Usage

LabelClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> RKernel::ValueWidget -> RKernel::StringWidget -> Textarea

Public fields

_model_name

Name of the Javascript model in the frontend

_view_name

Name of the Javascript view in the frontend

disabled

A logical value, whether the slider is disabled

continuous_update

A logical value, whether values should be updated as the slider is moved by the user

rows

An integer, the number of rows

style

The Text style, an object of class "TextStyleClass"

Methods

Public methods

TextareaClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::ValueWidget$initialize()

RKernel::ValueWidget$on_change()

Method clone()

The objects of this class are cloneable with this method.

Usage

TextareaClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> RKernel::ValueWidget -> RKernel::StringWidget -> TextWidget

Public fields

_model_name

Name of the Javascript model in the frontend

_view_name

Name of the Javascript view in the frontend

disabled

A logical value, whether the slider is disabled

continuous_update

A logical value, whether values should be updated as the slider is moved by the user

style

The Text style, an object of class "TextStyleClass"

Methods

Public methods

TextWidgetClass$clear()

TextWidgetClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::ValueWidget$initialize()

RKernel::ValueWidget$on_change()

Method clear()

Clear the text area

Usage

TextWidgetClass$clear()

Method clone()

The objects of this class are cloneable with this method.

Usage

TextWidgetClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> RKernel::ValueWidget -> RKernel::StringWidget -> RKernel::TextWidget -> PasswordWidget

Public fields

_model_name

Name of the Javascript model in the frontend

_view_name

Name of the Javascript view in the frontend

Methods

Public methods

PasswordWidgetClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::ValueWidget$initialize()

RKernel::ValueWidget$on_change()

RKernel::TextWidget$clear()

Method clone()

The objects of this class are cloneable with this method.

Usage

PasswordWidgetClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> RKernel::ValueWidget -> RKernel::StringWidget -> RKernel::TextWidget -> Combobox

Public fields

_model_name

Name of the Javascript model in the frontend

_view_name

Name of the Javascript view in the frontend

options

A unicode vector, the available options

ensure_option

A boolean (logical) value, whether at least one option has to be activated

Methods

Public methods

ComboboxClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::ValueWidget$initialize()

RKernel::ValueWidget$on_change()

RKernel::TextWidget$clear()

Method clone()

The objects of this class are cloneable with this method.

Usage

ComboboxClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.