Tags Input Widgets

Classes and constructors to great tag input widgets

Usage

TagsInput(...)

ColorsInput(...)

FloatsInput(...)

IntsInput(...)

Arguments

...

Arguments passed to the inializer

Functions

TagsInput(): A taginput constructor

ColorsInput(): A color taginput constructor

FloatsInput(): A color taginput constructor

IntsInput(): A color taginput constructor

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> RKernel::ValueWidget -> TagsInputBase

Public fields

_model_name

Name of the Javascript model in the frontend

value

A list of tags

placeholder

A placeholder string

allowed_tags

Optional list with allowed tags.

allow_duplicates

Logical, whether duplicate tags are allowed.

required_version

Minimum required ipywidgets version in which the current widget class is supported.

Methods

Public methods

TagsInputBaseClass$validate_value()

TagsInputBaseClass$new()

TagsInputBaseClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::ValueWidget$on_change()

Method validate_value()

Check value for validity,

Usage

TagsInputBaseClass$validate_value(value)

Arguments

value

A character string with one or more tags.

Method new()

Initializer function

Usage

TagsInputBaseClass$new(...)

Arguments

...

Arguments passed to the superclass initializer

Method clone()

The objects of this class are cloneable with this method.

Usage

TagsInputBaseClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> RKernel::ValueWidget -> RKernel::TagsInputBase -> TagsInput

Public fields

_model_name

Name of the Javascript model in the frontend

_view_name

Name of the Javascript view in the frontend.

value

A list of tags as unicode strings

tag_style

The string that describes the tag style

Methods

Public methods

TagsInputClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::ValueWidget$on_change()

RKernel::TagsInputBase$initialize()

RKernel::TagsInputBase$validate_value()

Method clone()

The objects of this class are cloneable with this method.

Usage

TagsInputClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> RKernel::ValueWidget -> RKernel::TagsInputBase -> ColorsInput

Public fields

_model_name

Name of the Javascript model in the frontend

_view_name

Name of the Javascript view in the frontend.

value

A list of tags as unicode strings

Methods

Public methods

ColorsInputClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::ValueWidget$on_change()

RKernel::TagsInputBase$initialize()

RKernel::TagsInputBase$validate_value()

Method clone()

The objects of this class are cloneable with this method.

Usage

ColorsInputClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> RKernel::ValueWidget -> RKernel::TagsInputBase -> RKernel::TagsInput -> ColorsInput

Public fields

_model_name

Name of the Javascript model in the frontend

min

A Float traitlet, the minimum allowed value.

max

A Float traitlet, the maximum allowed value.

value

A list of numeric tags

Methods

Public methods

NumbersInputBase$validate_value()

NumbersInputBase$new()

NumbersInputBase$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::ValueWidget$on_change()

Method validate_value()

Check value for validity,

Usage

NumbersInputBase$validate_value(value)

Arguments

value

A character string with one or more tags.

Method new()

Initializer function

Usage

NumbersInputBase$new(...)

Arguments

...

Arguments passed to the superclass initializer

value

An initial value

Method clone()

The objects of this class are cloneable with this method.

Usage

NumbersInputBase$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> RKernel::ValueWidget -> RKernel::TagsInputBase -> RKernel::TagsInput -> RKernel::ColorsInput -> FloatsInput

Public fields

_model_name

Name of the Javascript model in the frontend

_view_name

Name of the Javascript view in the frontend.

value

A list of numeric tags

format

The number format

Methods

Public methods

FloatsInputClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::ValueWidget$on_change()

RKernel::ColorsInput$initialize()

RKernel::ColorsInput$validate_value()

Method clone()

The objects of this class are cloneable with this method.

Usage

FloatsInputClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

RKernel::HasTraits -> RKernel::Widget -> RKernel::DOMWidget -> RKernel::DescriptionWidget -> RKernel::ValueWidget -> RKernel::TagsInputBase -> RKernel::TagsInput -> RKernel::ColorsInput -> IntsInput

Public fields

_model_name

Name of the Javascript model in the frontend

_view_name

Name of the Javascript view in the frontend.

min

A Integer traitlet, the minimum allowed value.

max

A Integer traitlet, the maximum allowed value.

value

A list of interger number tags

format

The number format

Methods

Public methods

IntsInputClass$clone()

Inherited methods

RKernel::HasTraits$notify()

RKernel::HasTraits$observe()

RKernel::HasTraits$validate()

RKernel::Widget$_send()

RKernel::Widget$check_version()

RKernel::Widget$close()

RKernel::Widget$display_data()

RKernel::Widget$finalize()

RKernel::Widget$get_state()

RKernel::Widget$handle_buffers()

RKernel::Widget$handle_comm_msg()

RKernel::Widget$handle_comm_opened()

RKernel::Widget$handle_custom_msg()

RKernel::Widget$handle_displayed()

RKernel::Widget$handle_event()

RKernel::Widget$on_displayed()

RKernel::Widget$on_event()

RKernel::Widget$on_msg()

RKernel::Widget$open()

RKernel::Widget$send()

RKernel::Widget$send_state()

RKernel::Widget$set_state()

RKernel::DOMWidget$add_class()

RKernel::DOMWidget$has_class()

RKernel::DOMWidget$remove_class()

RKernel::ValueWidget$on_change()

RKernel::ColorsInput$initialize()

RKernel::ColorsInput$validate_value()

Method clone()

The objects of this class are cloneable with this method.

Usage

IntsInputClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.